requirements.txt
python-dotenv>=1.0
openai>=1.51

httpx>=0.27,<0.29
feedparser>=6.0
dateparser>=1.1
pydantic>=2.12,<3
tenacity>=8.2
beautifulsoup4>=4.12
lxml>=5.3
selectolax>=0.3
PyMuPDF>=1.24

supabase>=2.23
gotrue>=2.12
postgrest>=2.23
realtime>=2.23
storage3>=2.23
websockets>=13,<16
